mcp-server/requirements.txt
httpx>=0.25.0
mcp-server>=0.1.4 